+++ CHANGES.rst
@@ -18,6 +18,10 @@ Major changes
 Minor changes
 -------------
 
+* The "Column name" column of the "summary statistics" table in the TableReport
+  is now always visible when scrolling the table. :pr:`1102` by :user:`Jérôme
+  Dockès <jeromedockes>`.
+
 Release 0.3.1
 =============